+++ selectrum.el
@@ -2477,20 +2477,30 @@ PREDICATE, see `read-file-name'."
          (completing-read-function
           (lambda (&rest args)
             (setq completing-read-function crf)
-            (when (and default-filename
-                       ;; ./ should be omitted.
-                       (not (equal
-                             (expand-file-name default-filename)
-                             (expand-file-name default-directory))))
-              (setf (nth 6 args)        ; DEFAULT
-                    ;; Sort for directories needs any final
-                    ;; slash removed.
-                    (directory-file-name
-                     ;; The candidate should be sorted by it's
-                     ;; relative name.
-                     (file-relative-name default-filename
-                                         default-directory))))
-            (apply #'selectrum--completing-read-file-name args))))
+            (let ((default (if (consp default-filename)
+                               (car default-filename)
+                             default-filename)))
+              ;; Get the default back for internal handling as it
+              ;; wasn't passed to `read-file-name-default'. See
+              ;; comment below.
+              (when default
+                (when (equal (file-name-directory
+                              (directory-file-name
+                               (expand-file-name
+                                default)))
+                             (file-name-directory
+                              (expand-file-name
+                               default-directory)))
+                  ;; Make the default sorted first by its relative name
+                  ;; when it is inside the prompting directory.
+                  (setq default
+                        (file-relative-name default default-directory)))
+                (if (consp default-filename)
+                    (setcar default-filename default)
+                  (setq default-filename default))
+                ;; Adjust the DEFAULT arg.
+                (setf (nth 6 args) default-filename))
+              (apply #'selectrum--completing-read-file-name args)))))
     (read-file-name-default
      prompt dir
      ;; We don't pass default-candidate here to avoid that